refactor: move viewer to microviewer and edit MANIFEST.in
