+++ AUTHORS
@@ -4,6 +4,7 @@ Manuel Castro <macastro@princeton.edu>
 Nicholas Turner <nturner.stanford@gmail.com>
 Nico Kemnitz <nkemnitz@princeton.edu>
 Pat Gunn <pgunn01@gmail.com>
+Shang Mu <smu@princeton.edu>
 Thomas Macrina <thomas.macrina@gmail.com>
 William Silversmith <william.silversmith@gmail.com>
 William Silversmith <ws9@princeton.edu>
