+++ MANIFEST.in
@@ -1,2 +1,1 @@
-recursive-include ext
 include LICENSE
